move a couple more funcs to localization, move test_locale to test_localization
